Dashboard: Extract route-agnostic backup components for reuse
Refactor the dashboard's backup UI into shared, route-agnostic pieces so
the list/detail/restore/download logic can be reused across variants:

* BackupsBrowser - list + detail grid + selection + activity-log fetch
* BackupRestoreFlow / BackupDownloadFlow - the restore/download step machines
* BackupDetails - restore/download navigation injected via callbacks
* BackupCardContent - overview card body parameterized by siteId + backupUrl

The dotcom backup pages (sites/backups, sites/backup-restore,
sites/backup-download) become thin shells over these components; behavior
is unchanged.
