+++ client/dashboard/sites/backups/index.tsx
@@ -3,32 +3,33 @@ import { siteBySlugQuery, siteSettingsQuery } from '@automattic/api-queries';
 import { DateRangePicker } from '@automattic/date-range-picker';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { Outlet, useParams, useRouter } from '@tanstack/react-router';
-import { __experimentalGrid as Grid } from '@wordpress/components';
-import { useViewportMatch } from '@wordpress/compose';
 import { useDispatch } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import { backup } from '@wordpress/icons';
 import { store as noticesStore } from '@wordpress/notices';
-import { useState, useEffect, useCallback } from 'react';
+import { useCallback } from 'react';
 import { FileBrowserProvider } from '../../../my-sites/backup/backup-contents-page/file-browser/file-browser-context';
 import Breadcrumbs from '../../app/breadcrumbs';
 import { useDateRange } from '../../app/hooks/use-date-range';
 import { useLocale } from '../../app/locale';
 import { PerformanceTrackerStop } from '../../app/performance-tracking';
-import { siteRoute, siteBackupsIndexRoute, siteBackupDetailRoute } from '../../app/router/sites';
-import { Card, CardBody } from '../../components/card';
+import {
+	siteRoute,
+	siteBackupsRoute,
+	siteBackupsIndexRoute,
+	siteBackupDetailRoute,
+	siteBackupRestoreRoute,
+	siteBackupDownloadRoute,
+} from '../../app/router/sites';
 import { PageHeader } from '../../components/page-header';
 import PageLayout from '../../components/page-layout';
 import { hasHostingFeature } from '../../utils/site-features';
 import HostingFeatureGatedWithCallout from '../hosting-feature-gated-with-callout';
 import { SitesNoticeArbiter } from '../notice-arbiter';
-import { BackupDetails } from './backup-details';
-import { BackupDetailsSkeleton } from './backup-details-skeleton';
 import { BackupNotices } from './backup-notices';
 import { BackupNowButton } from './backup-now-button';
+import { BackupsBrowser, useIsBackupsSmallViewport } from './backups-browser';
 import illustrationUrl from './backups-callout-illustration.svg';
-import { BackupsList } from './backups-list';
-import { useActivityLog } from './use-activity-log';
 import { useBackupState } from './use-backup-state';
 import './style.scss';
 import type { ActivityLogEntry } from '@automattic/api-core';
@@ -38,15 +39,13 @@ export function BackupsListPage() {
 	const { siteSlug } = siteRoute.useParams();
 	const router = useRouter();
 
-	const routeParams = useParams( {
-		strict: false,
-		shouldThrow: false,
-	} ) as { rewindId?: string } | undefined;
-
+	const routeParams = useParams( { strict: false, shouldThrow: false } ) as
+		| { rewindId?: string }
+		| undefined;
 	const rewindId = routeParams?.rewindId;
 
 	const { data: site } = useSuspenseQuery( siteBySlugQuery( siteSlug ) );
-
+	const searchParams = siteBackupsRoute.useSearch();
 	const backupState = useBackupState( site.ID );
 
 	const { data: siteSettings } = useSuspenseQuery( {
@@ -56,7 +55,6 @@ export function BackupsListPage() {
 			timezoneString: s?.timezone_string || undefined,
 		} ),
 	} );
-
 	const { gmtOffset, timezoneString } = siteSettings;
 
 	const { dateRange, handleDateRangeChange } = useDateRange( {
@@ -64,137 +62,56 @@ export function BackupsListPage() {
 		gmtOffset,
 		defaultDays: 30,
 	} );
-	const [ selectedBackup, setSelectedBackupInState ] = useState< ActivityLogEntry | null >( null );
 
-	const { activityLog, isLoadingActivityLog } = useActivityLog( {
-		siteId: site.ID,
-		dateRange,
-		gmtOffset,
-		timezoneString,
-	} );
-
-	const setSelectedBackup = useCallback(
-		( backup?: ActivityLogEntry | null, replace = false ) => {
-			if ( backup ) {
+	const navigateToBackup = useCallback(
+		( selected: ActivityLogEntry | null ) => {
+			if ( selected ) {
 				router.navigate( {
 					to: siteBackupDetailRoute.fullPath,
-					params: {
-						siteSlug,
-						rewindId: backup.rewind_id,
-					},
-					// The following preserves the existing query string.
+					params: { siteSlug, rewindId: selected.rewind_id },
 					search: ( query: Record< string, string > ) => query,
-					replace,
 				} );
 			} else {
 				router.navigate( {
 					to: siteBackupsIndexRoute.fullPath,
-					params: {
-						siteSlug,
-					},
-					// The following preserves the existing query string.
+					params: { siteSlug },
 					search: ( query: Record< string, string > ) => query,
-					replace,
 				} );
 			}
 		},
 		[ router, siteSlug ]
 	);
 
-	const isSmallViewport = useViewportMatch( 'xlarge', '<' );
+	const handleRequestRestore = ( selected: ActivityLogEntry ) => {
+		router.navigate( {
+			to: siteBackupRestoreRoute.fullPath,
+			params: { siteSlug, rewindId: selected.rewind_id },
+		} );
+	};
 
-	// Auto-select backup based on rewindId parameter
-	useEffect( () => {
-		if ( rewindId && activityLog ) {
-			const targetBackup = activityLog.find( ( item ) => item.rewind_id === rewindId );
-			if ( targetBackup ) {
-				setSelectedBackupInState( targetBackup );
-			}
-			return;
-		}
+	const handleRequestDownload = ( selected: ActivityLogEntry ) => {
+		router.navigate( {
+			to: siteBackupDownloadRoute.fullPath,
+			params: { siteSlug, rewindId: selected.rewind_id },
+		} );
+	};
 
-		// if no rewindId, then it's hitting the index route
-		// we select the first found backup without changing the route in that case to make things look nice.
-		// no selection if it's mobile!
-		const backup = activityLog?.[ 0 ];
-		if ( ! rewindId && backup && ! isSmallViewport ) {
-			setSelectedBackupInState( backup );
-		}
-
-		// no rewind id in param, and no backup? We have an empty query
-		// don't set any backup
-		if ( ! rewindId && ! backup ) {
-			setSelectedBackupInState( null );
-		}
-	}, [ rewindId, activityLog, setSelectedBackupInState, isSmallViewport ] );
+	const handleGranularDownloadReady = ( selected: ActivityLogEntry, downloadId: number ) => {
+		router.navigate( {
+			to: siteBackupDownloadRoute.fullPath,
+			params: { siteSlug, rewindId: selected.rewind_id },
+			search: { downloadId },
+		} );
+	};
 
 	const handleDateRangeChangeWrapper = ( next: { start: Date; end: Date } ) => {
 		handleDateRangeChange( next );
-		setSelectedBackup( null, false );
+		navigateToBackup( null );
 	};
-	const columns = isSmallViewport ? 1 : 2;
 
+	const isSmallViewport = useIsBackupsSmallViewport();
 	const hasBackups = hasHostingFeature( site, HostingFeatures.BACKUPS_SELF_SERVE );
-
-	const handleBackupSelection = ( backup: ActivityLogEntry | null ) => {
-		setSelectedBackup( backup );
-	};
-
-	const renderMobileView = () => {
-		if ( selectedBackup ) {
-			return (
-				<>
-					<PerformanceTrackerStop />
-					<BackupDetails
-						backup={ selectedBackup }
-						site={ site }
-						timezoneString={ timezoneString }
-						gmtOffset={ gmtOffset }
-					/>
-				</>
-			);
-		}
-
-		return (
-			<>
-				{ ! isLoadingActivityLog && <PerformanceTrackerStop /> }
-				<BackupsList
-					activityLog={ activityLog }
-					isLoadingActivityLog={ isLoadingActivityLog }
-					selectedBackup={ selectedBackup }
-					setSelectedBackup={ handleBackupSelection }
-					dateRange={ dateRange }
-					timezoneString={ timezoneString }
-					gmtOffset={ gmtOffset }
-				/>
-			</>
-		);
-	};
-
-	const renderDetailsPanel = () => {
-		if ( isLoadingActivityLog ) {
-			return <BackupDetailsSkeleton />;
-		}
-
-		if ( selectedBackup ) {
-			return (
-				<BackupDetails
-					backup={ selectedBackup }
-					site={ site }
-					timezoneString={ timezoneString }
-					gmtOffset={ gmtOffset }
-				/>
-			);
-		}
-
-		return (
-			<Card>
-				<CardBody style={ { minHeight: '300px' } } children={ null } />
-			</Card>
-		);
-	};
-
-	const isMobileDetailsView = isSmallViewport && selectedBackup;
+	const isMobileDetailsView = isSmallViewport && !! rewindId;
 	const shouldShowActions = hasBackups && ! isMobileDetailsView;
 
 	const actions = (
@@ -242,28 +159,22 @@ export function BackupsListPage() {
 				</>
 			}
 		>
-			{ hasBackups && (
-				<>
-					{ isSmallViewport ? (
-						renderMobileView()
-					) : (
-						<Grid columns={ columns } templateColumns="40% 1fr">
-							{ ! isLoadingActivityLog && <PerformanceTrackerStop /> }
-							<BackupsList
-								activityLog={ activityLog }
-								isLoadingActivityLog={ isLoadingActivityLog }
-								selectedBackup={ selectedBackup }
-								setSelectedBackup={ handleBackupSelection }
-								dateRange={ dateRange }
-								timezoneString={ timezoneString }
-								gmtOffset={ gmtOffset }
-							/>
-							{ renderDetailsPanel() }
-						</Grid>
-					) }
-				</>
+			{ hasBackups ? (
+				<BackupsBrowser
+					site={ site }
+					rewindId={ rewindId }
+					dateRange={ dateRange }
+					timezoneString={ timezoneString }
+					gmtOffset={ gmtOffset }
+					searchParams={ searchParams }
+					onSelectBackup={ navigateToBackup }
+					onRequestRestore={ handleRequestRestore }
+					onRequestDownload={ handleRequestDownload }
+					onGranularDownloadReady={ handleGranularDownloadReady }
+				/>
+			) : (
+				<PerformanceTrackerStop />
 			) }
-			{ ! hasBackups && <PerformanceTrackerStop /> }
 		</PageLayout>
 	);
 }
